tests/example/base-package/pyproject.toml
[project]
name = "base-package"
version = "0.1.0"
dependencies = ["dynapydantic"]

[project.scripts]
base-package = "base_package.cli:cli"

[project.entry-points."dynapydantic.plugins"]
base-package = "base_package"
